Ng2 Migration: Moving further ctrls to its own file, declaring global variables, updating code
