+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/controllers/ClaimCtrl.ts
@@ -1,3 +1,6 @@
+declare var getBaseUri: any;
+declare var orcidGA: any;
+
 import * as angular from 'angular';
 import {NgModule} from '@angular/core';
