+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/controllers/DeprecateAccountCtrl.ts
@@ -1,3 +1,7 @@
+declare var $: any;
+declare var colorbox: any;
+declare var getBaseUri: any;
+
 import * as angular from 'angular';
 import {NgModule} from '@angular/core';
