+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/controllers/PublicPeerReviewCtrl.ts
@@ -1,3 +1,6 @@
+declare var ActSortState: any;
+declare var GroupedActivities: any;
+
 import * as angular from 'angular';
 import {NgModule} from '@angular/core';
