+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/controllers/SecurityQuestionEditCtrl.ts
@@ -1,3 +1,8 @@
+declare var $: any;
+declare var colorbox: any;
+declare var getBaseUri: any;
+declare var orcidVar: any;
+
 import * as angular from 'angular';
 import {NgModule} from '@angular/core';
